move patch engine to pkg/patch
