+++ pkg/patch/apply.go
@@ -1,4 +1,4 @@
-package appsets
+package patch
 
 import (
 	"io"
