+++ pkg/patch/apply_test.go
@@ -1,4 +1,4 @@
-package appsets
+package patch
 
 import (
 	"os"
