+++ pkg/patch/loader.go
@@ -1,4 +1,4 @@
-package appsets
+package patch
 
 import (
 	"fmt"
